refactor(api): deprecate route POST /schooling-registration-dependent-users/recover-account
in favor of POST /sco-organization-learners/account-recovery
