--- api/lib/application/organization-learner-dependent-users/organization-learner-dependent-user-controller.js
@@ -1,14 +1,0 @@
-const usecases = require('../../domain/usecases');
-const studentInformationForAccountRecoverySerializer = require('../../infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer');
-
-module.exports = {
-  async checkScoAccountRecovery(request) {
-    const studentInformation = await studentInformationForAccountRecoverySerializer.deserialize(request.payload);
-
-    const studentInformationForAccountRecovery = await usecases.checkScoAccountRecovery({
-      studentInformation,
-    });
-
-    return studentInformationForAccountRecoverySerializer.serialize(studentInformationForAccountRecovery);
-  },
-};
