--- api/tests/unit/application/organization-learner-dependent-user/organization-learner-dependant-user-controller_test.js
@@ -1,51 +1,0 @@
-const { expect, sinon, hFake } = require('../../../test-helper');
-
-const usecases = require('../../../../lib/domain/usecases');
-const studentInformationForAccountRecoverySerializer = require('../../../../lib/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer.js');
-const organizationLearnerDependantUserController = require('../../../../lib/application/organization-learner-dependent-users/organization-learner-dependent-user-controller');
-
-describe('Unit | Application | Controller | organization-learner-user-associations', function () {
-  describe('#checkScoAccountRecovery', function () {
-    it('should return student account information serialized', async function () {
-      // given
-      const studentInformation = {
-        ineIna: '1234567890A',
-        firstName: 'Bob',
-        lastName: 'Camond',
-        birthdate: '2001-12-08',
-      };
-      const request = {
-        payload: {
-          data: {
-            type: 'student-information',
-            attributes: {
-              'ine-ina': studentInformation.ineIna,
-              'first-name': studentInformation.firstName,
-              'last-name': studentInformation.lastName,
-              birthdate: studentInformation.birthdate,
-            },
-          },
-        },
-      };
-      const studentInformationForAccountRecovery = Symbol();
-      const studentInformationForAccountRecoveryJSONAPI = Symbol();
-
-      sinon
-        .stub(studentInformationForAccountRecoverySerializer, 'deserialize')
-        .withArgs(request.payload)
-        .resolves(studentInformation);
-      sinon.stub(usecases, 'checkScoAccountRecovery');
-      usecases.checkScoAccountRecovery.withArgs({ studentInformation }).resolves(studentInformationForAccountRecovery);
-      sinon
-        .stub(studentInformationForAccountRecoverySerializer, 'serialize')
-        .withArgs(studentInformationForAccountRecovery)
-        .returns(studentInformationForAccountRecoveryJSONAPI);
-
-      // when
-      const response = await organizationLearnerDependantUserController.checkScoAccountRecovery(request);
-
-      // then
-      expect(response).to.deep.equal(studentInformationForAccountRecoveryJSONAPI);
-    });
-  });
-});
