+++ api/tests/unit/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -6,6 +6,7 @@ const scoOrganizationLearnerController = require('../../../../lib/application/sc
 const scoOrganizationLearnerSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/sco-organization-learner-serializer');
 const organizationLearnerUserAssociationSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/organization-learner-user-association-serializer');
 const organizationLearnerDependentUserSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/organization-learner-dependent-user-serializer');
+const studentInformationForAccountRecoverySerializer = require('../../../../lib/infrastructure/serializers/jsonapi/student-information-for-account-recovery-serializer');
 
 describe('Unit | Application | Controller | sco-organization-learner', function () {
   describe('#reconcileScoOrganizationLearnerManually', function () {
@@ -328,6 +329,86 @@ describe('Unit | Application | Controller | sco-organization-learner', function 
         path: '/api/sco-organization-learners/username-password-generation',
       };
       const response = await scoOrganizationLearnerController.generateUsernameWithTemporaryPassword(request, hFake);
+
+      // then
+      expect(response.headers['Deprecation']).to.not.exist;
+    });
+  });
+
+  describe('#checkScoAccountRecovery', function () {
+    const userId = 2;
+    let request = {
+      auth: { credentials: { userId } },
+      payload: { data: { attributes: {} } },
+    };
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'checkScoAccountRecovery');
+      usecases.checkScoAccountRecovery.resolves();
+      sinon.stub(studentInformationForAccountRecoverySerializer, 'serialize');
+      studentInformationForAccountRecoverySerializer.serialize.resolves();
+      sinon.stub(studentInformationForAccountRecoverySerializer, 'deserialize');
+      studentInformationForAccountRecoverySerializer.deserialize.resolves();
+    });
+
+    it('should return student account information serialized', async function () {
+      // given
+      hFake.request = { path: {} };
+      const studentInformation = {
+        ineIna: '1234567890A',
+        firstName: 'Bob',
+        lastName: 'Camond',
+        birthdate: '2001-12-08',
+      };
+      request = {
+        payload: {
+          data: {
+            type: 'student-information',
+            attributes: {
+              'ine-ina': studentInformation.ineIna,
+              'first-name': studentInformation.firstName,
+              'last-name': studentInformation.lastName,
+              birthdate: studentInformation.birthdate,
+            },
+          },
+        },
+      };
+      const studentInformationForAccountRecovery = Symbol();
+      const studentInformationForAccountRecoveryJSONAPI = Symbol();
+
+      studentInformationForAccountRecoverySerializer.deserialize.withArgs(request.payload).resolves(studentInformation);
+      usecases.checkScoAccountRecovery.withArgs({ studentInformation }).resolves(studentInformationForAccountRecovery);
+      studentInformationForAccountRecoverySerializer.serialize
+        .withArgs(studentInformationForAccountRecovery)
+        .returns(studentInformationForAccountRecoveryJSONAPI);
+
+      // when
+      const response = await scoOrganizationLearnerController.checkScoAccountRecovery(request, hFake);
+
+      // then
+      expect(response.source).to.deep.equal(studentInformationForAccountRecoveryJSONAPI);
+    });
+
+    it('should return information about deprecation when old route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/schooling-registration-dependent-users/recover-account',
+      };
+      const response = await scoOrganizationLearnerController.checkScoAccountRecovery(request, hFake);
+
+      // then
+      expect(response.headers['Deprecation']).to.equal('true');
+      expect(response.headers['Link']).to.equal(
+        '/api/sco-organization-learners/account-recovery; rel="successor-version"'
+      );
+    });
+
+    it('should not return information about deprecation when new route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/sco-organization-learners/account-recovery',
+      };
+      const response = await scoOrganizationLearnerController.checkScoAccountRecovery(request, hFake);
 
       // then
       expect(response.headers['Deprecation']).to.not.exist;
